+++ meilisearch-http/Cargo.toml
@@ -47,6 +47,7 @@ whoami = "0.8.1"
 http-service = "0.4.0"
 http-service-mock = "0.4.0"
 tempdir = "0.3.7"
+once_cell = "1.3.1"
 
 [dev-dependencies.assert-json-diff]
 git = "https://github.com/qdequele/assert-json-diff"
